test: consolidate webhook verify test into webhooks.spec.ts
Drop the file-wide `standardwebhooks` mock (only the verify test used it) and
rewrite the verify test as a real sign + verify round-trip, plus a bad-signature
case. Removes the separate webhooks-verify.spec.ts.
